Reading date from API with long type

+++ BadTracker/app/src/main/java/fr/efrei/badtracker/api/dtos/CreateMatchDto.java
@@ -5,9 +5,9 @@ import android.location.Location;
 public class CreateMatchDto {
     private final String name;
     private Location location;
-    private final int date;
+    private final long date;
 
-    public CreateMatchDto(String name, Location Location, int date) {
+    public CreateMatchDto(String name, Location Location, long date) {
         this.name = name;
         this.location = location;
         this.date = date;
